refactor(rebuild): extract trigger helpers + Phase 0 daily-sweep + Phase 1.5 prevention
Foundation work for the Trigger.dev → local-scheduler migration. Three
discrete pieces in one commit because they're tightly interrelated and
the prior branch (#467) was overwritten by an unrelated frontend merge.

# Helpers extracted to @/lib (deletable shims left at old paths)

- src/lib/telegram.ts (new) — pure Node, callable from any context
- src/lib/scraper-verification.ts (new) — moved from src/trigger/, swapped
  Gemini → DeepSeek-V4-Flash. Drops the old responseMimeType/JSON-schema
  options (DeepSeek's generateText doesn't expose them); existing
  stripCodeFences + JSON.parse fallback handles fenced or unfenced JSON.
  Timeout 10s → 15s for DeepSeek cold-cache margin.
- src/trigger/utils/telegram.ts → 1-line re-export shim
- src/trigger/verification.ts → 1-line re-export shim
- src/autoresearch/experiment-log.ts → import from @/lib/telegram directly

# Phase 0: extract daily-sweep into pure-Node job module

- src/lib/jobs/daily-sweep.ts (new, ~430 LOC) — exports runDailySweep()
  with all 5 phases (TMDB matching, backfill, Letterboxd, posters, data
  quality). Skips Mondays by default; opts in via skipMondayGuard.
- src/trigger/enrichment/daily-sweep.ts → 13-line shim around runDailySweep.

This makes daily-sweep callable from the upcoming Bree scheduler, admin
API routes, or CLI without any Trigger.dev runtime dependency.

# Phase 1.5: prevention layer in scrapers/pipeline.ts

Synthesised from 1,363 /data-check patrol logs in
Obsidian Vault/Pictures/Data Quality/. Top recurring issue across 6
weeks: anniversary / re-release / restoration suffixes generating
duplicate film rows. A single 2026-04-26 patrol cycle merged 22 dupes
of "Amélie", 9 of "Dr. Strangelove", recurring Akira.

normalizeTitle now calls cleanFilmTitle first. Both
"Amélie - 25th Anniversary" and "Amélie" produce key "amelie", so the
second scrape hits the existing film instead of creating a dupe.

Tests: 9 new prevention cases covering anniversary suffixes,
re-release/restoration suffixes, HTML entities, Latin-1 mojibake from
real patrol-log examples, idempotence, and false-positive guard for
legitimate titles containing "Anniversary". 110/110 tests pass
(85 pipeline + 25 agents).

No data migration needed — existing duplicates remain in the DB and
will be merged by the data-check patrol; this fix prevents new ones
from being created.
